+++ src/management/DeviceCredentialsManager.js
@@ -1,7 +1,7 @@
-var RestClient = require('rest-facade').Client;
 var ArgumentError = require('rest-facade').ArgumentError;
 var utils = require('../utils');
-
+var Auth0RestClient = require('../Auth0RestClient');
+var RetryRestClient = require('../RetryRestClient');
 
 /**
  * Simple facade for consuming a REST API endpoint.
@@ -19,6 +19,7 @@ var utils = require('../utils');
  * @param {Object} options            The client options.
  * @param {String} options.baseUrl    The URL of the API.
  * @param {Object} [options.headers]  Headers to be included in all requests.
+ * @param {Object} [options.retry]    Retry Policy Config
  */
 var DeviceCredentialsManager = function (options) {
   if (options === null || typeof options !== 'object') {
@@ -49,9 +50,10 @@ var DeviceCredentialsManager = function (options) {
    * {@link https://auth0.com/docs/api/v2#!/Device_Credentials
    *  Auth0 DeviceCredentialsManagers endpoint}.
    *
-   * @type {external:RestDeviceCredentialsManager}
+   * @type {external:RestClient}
    */
-  this.resource = new RestClient(options.baseUrl + '/device-credentials/:id', clientOptions);
+  var auth0RestClient = new Auth0RestClient(options.baseUrl + '/device-credentials/:id', clientOptions, options.tokenProvider);
+  this.resource = new RetryRestClient(auth0RestClient, options.retry);
 };